closing connections are now separated by operative system

+++ close_all.py
@@ -18,19 +18,15 @@ sys.path.append(os.path.join(BASE_DIR, 'lib'))
 from lib_log_nbz import Logging
 logger = Logging()
 
+if os.name == 'posix':
+    ppid = os.getppid()
+    logs = ['bmp.log', 'geckodriver.log', 'server.log']
+    for log in logs:
+        if os.path.isfile(os.path.join(os.getcwd(), log)):
+                os.remove(os.path.join(os.getcwd(), log))
+    logger.log('NOTE', 'Connections closed')
+    os.killpg(ppid, 9)
 
-for process in psutil.process_iter():
-    try:
-        process_info = process.as_dict(attrs=['name', 'cmdline'])
-        if process_info.get('name') in ('java', 'java.exe', 'Xvfb'):
-            for cmd_info in process_info.get('cmdline'):
-                if cmd_info == '-Dapp.name=browsermob-proxy' or cmd_info == ':99':
-                    if process.ppid() == psutil.Process(os.getpid()).ppid() or psutil.Process(process.ppid()).ppid() == 1:
-                        process.kill()
-    except psutil.NoSuchProcess:
-        pass
-
-logs = ['geckodriver.log', 'bmp.log', 'server.log']
-for log in logs:
-    if os.path.isfile(os.path.join(os.getcwd(), log)):
-        os.remove(os.path.join(os.getcwd(), log))
+elif os.name == 'nt':
+    #TODO all
+    pass
